feat(grading): Enforce whole-number score overrides in Exam Results
- Implement validation to ensure scores are whole numbers and do not exceed max marks.
- Update frontend to handle score input and display appropriately.
- Enhance backend to enforce score constraints for grading updates.

+++ backend/schemas/grade.py
@@ -19,7 +19,7 @@ class GradeResponse(BaseModel):
 
 
 class GradeUpdate(BaseModel):
-    score: float = Field(..., ge=0)
+    score: int = Field(..., ge=0)
     feedback: Optional[str] = None
